Added functionality and README
I was able to add functionality that lets you select colors of individual trajectories. Also, did some slight modifications that streamlined it. I was able to find the turtle object or rather dictionary and alter the elements to
 1. Remove the turtle graphic, making it
     look much cleaner
2. Bolster the thickness of the line,
    making it more visible.

I added a README to end off this bit of programming. Next thing on the docket is figuring out how to graph the data.